make all spotify object model keys able to be nil
While Spotify API sometimes explicitly says that a type can be nil,
there are sometimes implicit things that can cause object keys to
not be returned (nil as the api sees it).  This includes differences
in scopes/permissions and more.  To make this non-guarantee explicit
I have made all the types nil-able in their perspective object models.

The types shared between object models (the ones that are defined in
base modules like Spotify.Albums or Spotify.Artists) were not changed
to all be nil-able, keeping what their perspective documentation says.
Some of those types were already explicitly nil-able.

This was done in an effort to hope that users of the library will
recognize that any return value is not guaranteed and to make sure
to check/pattern-match the api results to not fall into nil values.